refactor(capreg): split 896-line sqlite_store.go by concern (core→94)
Pure same-package relocation, no call-site or behavior changes:

- sqlite_store.go (94): store struct, ctor, migrate entrypoint, Close
- sqlite_schema.go: V8 schema DDL + V1..V7→V8 migrations + column backfill
- sqlite_ops.go: Remember/Lookup/Record store operations
- sqlite_scan.go: row scanning + device/network context extraction

27 funcs in, 27 funcs out (lossless). Verified by go build, go vet,
the capreg schema+store tests, scoped golangci-lint, goimports.
